moves bad context loading out of the crossdaolistcard
